Modify string-rewriting code to handle edge case.
When MLton and PolyML encounter a string "hello\   \^Hworld", it is rewritten as "hello^Hworld". The skipping of formatting characters is rewritten instead of the \^H control code.

Previously, this lexer did not handle that properly, rewriting the control code instead of skipping the formatting characters. That is fixed now, so it is rewritten to give the same result as MLton and PolyML.